--- a/h2o-core/src/main/java/water/rapids/RadixOrder.java
+++ b/h2o-core/src/main/java/water/rapids/RadixOrder.java
@@ -4,6 +4,7 @@ import water.H2O;
 import water.Key;
 import water.MRTask;
 import water.RPC;
+import water.fvec.Chunk;
 import water.fvec.Frame;
 import water.fvec.Vec;
 import water.util.ArrayUtils;
@@ -147,8 +148,12 @@ class RadixOrder extends H2O.H2OCountedCompleter<RadixOrder> {
         double colMin = col.min();
         double colMax = col.max();
         if (col.isInt()) {
-          _base[i] = BigInteger.valueOf(Math.min((long)colMin, (long)colMax*(_ascending[i])));
-          max = BigInteger.valueOf(Math.max((long)colMax, (long)colMin*(_ascending[i])));
+          GetLongStatsTask glst = GetLongStatsTask.getLongStats(col);
+          long colMini = glst._colMin;
+          long colMaxi = glst._colMax;
+
+          _base[i] = BigInteger.valueOf(Math.min(colMini, colMaxi*(_ascending[i])));
+          max = BigInteger.valueOf(Math.max(colMaxi, colMini*(_ascending[i])));
         } else{
           _base[i] = MathUtils.convertDouble2BigInteger(Math.min(col.min(), colMax*(_ascending[i])));
           max = MathUtils.convertDouble2BigInteger(Math.max(col.max(), colMin*(_ascending[i])));
@@ -170,6 +175,25 @@ class RadixOrder extends H2O.H2OCountedCompleter<RadixOrder> {
 
       _bytesUsed[i] = Math.min(8, (_shift[i]+15) / 8);  // should not go over 8 bytes
       //assert (biggestBit-1)/8 + 1 == _bytesUsed[i];
+    }
+  }
+
+  // TODO: push these into Rollups?
+  private static class GetLongStatsTask extends MRTask<GetLongStatsTask> {
+    long _colMin=Long.MAX_VALUE;
+    long _colMax=Long.MIN_VALUE;
+    static GetLongStatsTask getLongStats(Vec col) {
+      return new GetLongStatsTask().doAll(col);
+    }
+    @Override public void map(Chunk c) {
+      for(int i=0; i<c._len; ++i) {
+        _colMin = Math.min(_colMin, c.at8(i));
+        _colMax = Math.max(_colMax, c.at8(i));
+      }
+    }
+    @Override public void reduce(GetLongStatsTask that) {
+      _colMin = Math.min(_colMin, that._colMin);
+      _colMax = Math.max(_colMax, that._colMax);
     }
   }