--- a/h2o-core/src/test/java/water/fvec/FrameTest.java
+++ b/h2o-core/src/test/java/water/fvec/FrameTest.java
@@ -11,8 +11,6 @@ import java.util.Set;
 
 import static org.junit.Assert.*;
 
-import static org.junit.Assert.assertTrue;
-
 /**
  * Tests for Frame.java
  */