fix lossy conversions of longs to doubles:
   - add at8_impl's for (C1S/C2S)Chunk
   - add new test cases to (C1S/C2S)ChunkTest
   - use BigInteger and smarter heueristics to detect "long" overflow