@wendycwong 2018-06-13
MichalK code review comments:
1. Remove operation with BigInteger
2. add Jmh test, check ChunkTest