@wendycwong 2018-06-26
Final timing test with Master and this branch only:


  | Spencer+Wendy | Master | % speedup by Spencer+Wendy
-- | -- | -- | --
Float | 7.49 | 10.74 | 43.28
Integer | 3.04 | 3.12 | 2.92
Long | 2.07 | 3.18 | 53.14
Double | 2.66 | 3.60 | 35.16
BigLong | 3.21 | 3.24 | 0.91